+++ Starstorm 2/Modules/EntityStates/BaseMeleeAttack.cs
@@ -47,9 +47,13 @@ namespace EntityStates.SS2UStates.Common
         private BaseState.HitStopCachedState hitStopCachedState;
         private Vector3 storedVelocity;
 
+        private float lastUpdateTime;
+
         public override void OnEnter()
         {
             base.OnEnter();
+
+            lastUpdateTime = Time.time;
 
             this.startedSkillStationary = (base.inputBank && base.inputBank.moveVector == Vector3.zero);
 
@@ -161,11 +165,13 @@ namespace EntityStates.SS2UStates.Common
         public override void FixedUpdate()
         {
             base.FixedUpdate();
+            float deltaTime = Time.time - lastUpdateTime;
+            lastUpdateTime = Time.time;
 
-            this.hitPauseTimer -= Time.fixedDeltaTime;
+            this.hitPauseTimer -= deltaTime;
             if (!this.inHitPause)
             {
-                this.stopwatch += Time.fixedDeltaTime;
+                this.stopwatch += deltaTime;
             }
 
             if (base.isAuthority)