+++ Starstorm 2/Survivors/Chirr/EntityStates/ChirrMain.cs
@@ -26,6 +26,13 @@ namespace EntityStates.SS2UStates.Chirr
         private bool inJetpackState = false;
         private EntityStateMachine jetpackStateMachine;
 
+        //Copied fron new artificer code
+        public bool jumpButtonState;
+        private bool heldPress;
+        private bool jumpToggledState;
+        private float oldJumpHeldTime;
+        private float jumpButtonHeldTime;
+
         public override void OnEnter()
         {
             base.OnEnter();
@@ -63,16 +70,66 @@ namespace EntityStates.SS2UStates.Chirr
             inJetpackState = this.jetpackStateMachine.state.GetType() == typeof(JetpackOn);
             if (this.hasCharacterMotor && this.hasInputBank && base.isAuthority)
             {
-                bool inputPressed = base.inputBank.jump.down && base.characterMotor.velocity.y < 0f && !base.characterMotor.isGrounded;
-                if (inputPressed && !inJetpackState)
+                NetworkUser networkUser = NetworkUser.readOnlyLocalPlayersList[0];
+                bool? flag;
+                if (networkUser == null)
+                {
+                    flag = null;
+                }
+                else
+                {
+                    LocalUser localUser = networkUser.localUser;
+                    flag = ((localUser != null) ? new bool?(localUser.userProfile.toggleArtificerHover) : null);
+                }
+                if (flag ?? true)
+                {
+                    if (base.inputBank.jump.down)
+                    {
+                        this.oldJumpHeldTime = this.jumpButtonHeldTime;
+                        this.jumpButtonHeldTime += Time.deltaTime;
+                        this.heldPress = (this.oldJumpHeldTime < 0.5f && this.jumpButtonHeldTime >= 0.5f);
+                    }
+                    else
+                    {
+                        this.oldJumpHeldTime = 0f;
+                        this.jumpButtonHeldTime = 0f;
+                        this.heldPress = false;
+                    }
+                    if (!base.characterMotor.isGrounded)
+                    {
+                        if (base.characterMotor.jumpCount == base.characterBody.maxJumpCount)
+                        {
+                            if (base.inputBank.jump.justPressed)
+                            {
+                                this.jumpButtonState = !this.jumpButtonState;
+                            }
+                        }
+                        else if (this.heldPress)
+                        {
+                            this.jumpButtonState = !this.jumpButtonState;
+                        }
+                    }
+                    else
+                    {
+                        this.jumpButtonState = false;
+                    }
+                }
+                else
+                {
+                    this.jumpButtonState = base.inputBank.jump.down;
+                }
+                bool shouldJetpack = this.jumpButtonState && base.characterMotor.velocity.y < 0f && !base.characterMotor.isGrounded;
+                bool isInJetpack = this.jetpackStateMachine.state.GetType() == typeof(JetpackOn);
+                if (shouldJetpack && !isInJetpack)
                 {
                     this.jetpackStateMachine.SetNextState(new JetpackOn());
                 }
-                if (inJetpackState && !inputPressed)
+                if (!shouldJetpack && isInJetpack)
                 {
                     this.jetpackStateMachine.SetNextState(new Idle());
                 }
             }
+            base.ProcessJump();
         }
 
         public override void FixedUpdate()