+++ Starstorm 2/Survivors/Chirr/EntityStates/JetpackOn.cs
@@ -4,19 +4,25 @@ namespace EntityStates.SS2UStates.Chirr
 {
 	public class JetpackOn : BaseState
 	{
+		private float lastUpdateTime;
         public override void OnEnter()
         {
             base.OnEnter();
+			lastUpdateTime = Time.time;
 			base.PlayCrossfade("Jetpack, Override", "Sprint", 0.5f);
         }
 
         public override void FixedUpdate()
 		{
 			base.FixedUpdate();
+
+			float deltaTime = Time.time - lastUpdateTime;
+			lastUpdateTime = Time.time;
+
 			if (base.isAuthority)
 			{
 				float num = base.characterMotor.velocity.y;
-				num = Mathf.MoveTowards(num, JetpackOn.hoverVelocity, JetpackOn.hoverAcceleration * Time.fixedDeltaTime);
+				num = Mathf.MoveTowards(num, JetpackOn.hoverVelocity, JetpackOn.hoverAcceleration * deltaTime);
 				base.characterMotor.velocity = new Vector3(base.characterMotor.velocity.x, num, base.characterMotor.velocity.z);
 
                 if (base.characterBody && !base.characterBody.isSprinting && base.inputBank)